requirements.txt
# Discord
discord.py[voice]>=2.0.0
PyNaCl>=1.5.0

# YouTube
yt-dlp>=2024.0.0

# Spotify
spotipy>=2.23.0

# Utilitários
python-dotenv>=1.0.0
